Reformatted Gui to be imported into Main
I added a a function to the submit button to throw all of the values into a dictionary and send it off to the main file. There is also some code to import the statistics from the main file and display them in the text box on the Gui. Finally, I changed the naming of some things to PEP8 better